requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.2
